cedar: minor renames
Addresses IDX-142
