+++ policy_set_test.go
@@ -13,17 +13,17 @@ func TestNewPolicySetFromFile(t *testing.T) {
 	t.Parallel()
 	t.Run("err-in-tokenize", func(t *testing.T) {
 		t.Parallel()
-		_, err := cedar.NewPolicySetFromFile("policy.cedar", []byte(`"`))
+		_, err := cedar.NewPolicySetFromBytes("policy.cedar", []byte(`"`))
 		testutil.Error(t, err)
 	})
 	t.Run("err-in-parse", func(t *testing.T) {
 		t.Parallel()
-		_, err := cedar.NewPolicySetFromFile("policy.cedar", []byte(`err`))
+		_, err := cedar.NewPolicySetFromBytes("policy.cedar", []byte(`err`))
 		testutil.Error(t, err)
 	})
 	t.Run("annotations", func(t *testing.T) {
 		t.Parallel()
-		ps, err := cedar.NewPolicySetFromFile("policy.cedar", []byte(`@key("value") permit (principal, action, resource);`))
+		ps, err := cedar.NewPolicySetFromBytes("policy.cedar", []byte(`@key("value") permit (principal, action, resource);`))
 		testutil.OK(t, err)
 		testutil.Equals(t, ps.GetPolicy("policy0").Annotations(), cedar.Annotations{"key": "value"})
 	})
@@ -162,7 +162,7 @@ forbid (
 
 	ps := cedar.NewPolicySet()
 	for i, p := range policies {
-		p.SetSourceFile("example.cedar")
+		p.SetFileName("example.cedar")
 		ps.UpsertPolicy(cedar.PolicyID(fmt.Sprintf("policy%d", i)), p)
 	}
