+++ policy_test.go
@@ -94,25 +94,3 @@ func TestPolicyAST(t *testing.T) {
 
 	_ = cedar.NewPolicyFromAST(astExample)
 }
-
-func TestPolicySlice(t *testing.T) {
-	t.Parallel()
-
-	policiesStr := `permit (
-    principal,
-    action == Action::"editPhoto",
-    resource
-)
-when { resource.owner == principal };
-
-forbid (
-    principal in Groups::"bannedUsers",
-    action,
-    resource
-);`
-
-	var policies cedar.PolicySlice
-	testutil.OK(t, policies.UnmarshalCedar([]byte(policiesStr)))
-
-	testutil.Equals(t, string(policies.MarshalCedar()), policiesStr)
-}
